requirements.txt
cymysql==0.9.12
Django==2.1.2
django-cors-headers==2.4.0
django-cymysql==2.1.0
djangorestframework==3.8.2
pytz==2018.5
